[eth] changed private key to be value object and removed factory method
