+++ src/Utils.ts
@@ -1,0 +1,16 @@
+import {randomBytes} from 'crypto';
+import {PrivateKey} from './ValueObjects';
+
+export default class Utils {
+
+    /**
+     *
+     * @returns {PrivateKey}
+     */
+    public manufakturPrivKey() : PrivateKey {
+
+        return new PrivateKey(randomBytes(32));
+
+    }
+
+}
