+++ test/EthWallet/EthUtils.spec.ts
@@ -1,13 +1,13 @@
 import {EthUtils} from "../../src/EthWallet/EthUtils";
 import {SecureStorageInterface} from "BITNATION-Panthalassa-TS-secure-storage-interface/SecureStorageInterface";
 import {instance, mock, when} from "ts-mockito";
-import PrivateKey from "../../src/EthWallet/PrivateKey";
 import {AES, enc} from 'crypto-js'
 import {mnemonicToEntropy, entropyToMnemonic, validateMnemonic} from 'bip39'
 import {unlinkSync} from "fs";
 import {isValidPrivate, isValidAddress} from "ethereumjs-util";
 import {EthKeyAlreadyExist, EthKeyDoesNotExist} from "../../src/Errors";
 import {NodeJsSecureStorage} from "BITNATION-Panthalassa-TS-node-js-secure-storage/NodeJsSecureStorage"
+import Utils from "../../src/Utils";
 
 const PRIVATE_TEST_KEY = "9b4b3da7d2d1c8749743a4ac17c151405f7900832bb4cf88735721cef4627096";
 
@@ -30,7 +30,7 @@ describe('EthUtils', () => {
             const ssMockInstance = new SecureStorageMock();
 
             //Eth utils
-            const eu:EthUtils = new EthUtils(ssMockInstance, PrivateKey);
+            const eu:EthUtils = new EthUtils(ssMockInstance, new Utils);
 
             //Should throw an error
             let error:any;
@@ -56,7 +56,7 @@ describe('EthUtils', () => {
             secureStorage.init();
 
             //Eth utils
-            const eu:EthUtils = new EthUtils(secureStorage, PrivateKey);
+            const eu:EthUtils = new EthUtils(secureStorage, new Utils);
 
             //Create keypair
             const keyPair = await eu.createEthKeyPair('my_secret_password');
@@ -86,7 +86,7 @@ describe('EthUtils', () => {
             secureStorage.init();
 
             //Eth utils
-            const eu:EthUtils = new EthUtils(secureStorage, PrivateKey);
+            const eu:EthUtils = new EthUtils(secureStorage, new Utils());
 
             const keyPair:{address: string, privKeyMnemonic: string} = await eu.createEthKeyPair('pwd');
 
@@ -110,7 +110,7 @@ describe('EthUtils', () => {
             when(secureStorageMock.hasItem(EthUtils.PRIV_KEY_SS_NAME)).thenReturn(new Promise(((resolve, reject) => resolve(true))));
 
             //Eth Utils
-            const eu:EthUtils = new EthUtils(instance(secureStorageMock), PrivateKey);
+            const eu:EthUtils = new EthUtils(instance(secureStorageMock), new Utils);
 
             //Should have an item
             expect(await eu.hasPrivKey()).toBeTruthy();
@@ -126,7 +126,7 @@ describe('EthUtils', () => {
             when(secureStorageMock.hasItem(EthUtils.PRIV_KEY_SS_NAME)).thenReturn(new Promise(((resolve, reject) => resolve(false))));
 
             //Eth Utils
-            const eu:EthUtils = new EthUtils(instance(secureStorageMock), PrivateKey);
+            const eu:EthUtils = new EthUtils(instance(secureStorageMock), new Utils);
 
             //Should not have an item
             expect(await eu.hasPrivKey()).toBeFalsy();
@@ -146,7 +146,7 @@ describe('EthUtils', () => {
             when(secureStorageMock.hasItem(EthUtils.PRIV_KEY_SS_NAME)).thenReturn(new Promise(((resolve, reject) => resolve(false))));
 
             //Eth Utils
-            const eu:EthUtils = new EthUtils(instance(secureStorageMock), PrivateKey);
+            const eu:EthUtils = new EthUtils(instance(secureStorageMock), new Utils);
 
             //Should throw an error
             let error:any;
@@ -186,7 +186,7 @@ describe('EthUtils', () => {
             when(secureStorageMock.getItem(EthUtils.PRIV_KEY_SS_NAME)).thenReturn(new Promise(((resolve, reject) => resolve(PRIVATE_KEY_ENCRYPTED))));
 
             //Eth Utils
-            const eu:EthUtils = new EthUtils(instance(secureStorageMock), PrivateKey);
+            const eu:EthUtils = new EthUtils(instance(secureStorageMock), new Utils);
 
             //Fetch private key
             const privKey:{privKey: string, privKeyMnemonic: string} = await eu.getPrivKey("my_secret_password");
@@ -212,7 +212,7 @@ describe('EthUtils', () => {
             ss.init();
 
             //Eth Utils
-            const ethUtils = new EthUtils(ss, PrivateKey);
+            const ethUtils = new EthUtils(ss, new Utils);
             await ethUtils.createEthKeyPair('passwd');
 
             //get address and assert it
